+++ common/src/utils/network/http/qnbytearrayref.cpp
@@ -79,7 +79,7 @@ const QnByteArrayConstRef::value_type& QnByteArrayConstRef::operator[]( size_typ
 
 QnByteArrayConstRef::operator QByteArray() const
 {
-    return m_src->mid( m_offset, m_count );
+    return m_src->mid( m_offset, (int) m_count );
 }
